@@ -3,9 +3,9 @@
 	"private": true,
 	"type": "module",
 	"scripts": {
-		"dev": "vite dev --port 3000",
-		"build": "vite build",
-		"start": "node .output/server/index.mjs",
+		"dev": "vite dev --port 3000 --host",
+		"build": "bunx prisma generate && vite build",
+		"start": ".output/server/index.mjs",
 		"preview": "vite preview",
 		"test": "vitest run",
 		"db:generate": "bunx prisma generate",
@@ -15,51 +15,58 @@
 		"lint": "biome check"
 	},
 	"dependencies": {
-		"@tanstack/react-devtools": "^0.9.5",
-		"@tanstack/react-router": "^1.159.5",
-		"@tanstack/react-router-devtools": "^1.159.5",
-		"@tanstack/react-router-ssr-query": "^1.159.5",
-		"@tanstack/react-start": "^1.159.5",
-		"@tanstack/react-table": "^8.21.3",
-		"@tanstack/router-plugin": "^1.159.5",
-		"@base-ui/react": "^1.1.0",
-		"@fontsource-variable/jetbrains-mono": "^5.2.8",
-		"@polar-sh/sdk": "^0.43.1",
-		"@prisma/adapter-pg": "^7.4.0",
-		"@prisma/client": "^7.4.0",
-		"@tailwindcss/vite": "^4.0.6",
-		"better-auth": "^1.4.18",
-		"class-variance-authority": "^0.7.1",
-		"clsx": "^2.1.1",
-		"framer-motion": "^12.4.3",
-		"gsap": "^3.12.5",
-		"@gsap/react": "^2.1.2",
-		"lucide-react": "^0.563.0",
+		"@base-ui/react": "1.2.0",
+		"@fontsource-variable/jetbrains-mono": "5.2.8",
+		"@gsap/react": "2.1.2",
+		"@polar-sh/sdk": "0.43.1",
+		"@prisma/adapter-pg": "7.4.0",
+		"@prisma/client": "7.4.0",
+		"@prisma/client-runtime-utils": "7.4.0",
+		"@tailwindcss/vite": "4.1.18",
+		"@tanstack/react-devtools": "0.9.5",
+		"@tanstack/react-router": "1.160.2",
+		"@tanstack/react-router-devtools": "1.160.2",
+		"@tanstack/react-router-ssr-query": "1.160.2",
+		"@tanstack/react-start": "1.160.2",
+		"@tanstack/react-table": "8.21.3",
+		"@tanstack/router-plugin": "1.160.2",
+		"better-auth": "1.4.18",
+		"class-variance-authority": "0.7.1",
+		"clsx": "2.1.1",
+		"framer-motion": "12.34.1",
+		"gsap": "3.14.2",
+		"lucide-react": "0.563.0",
 		"nitro": "latest",
-		"pg": "^8.13.1",
-		"react": "^19.2.0",
-		"react-dom": "^19.2.0",
-		"react-intersection-observer": "^9.15.1",
-		"shadcn": "^3.8.3",
-		"tailwind-merge": "^3.4.0",
-		"tailwindcss": "^4.0.6",
-		"tw-animate-css": "^1.4.0",
-		"vite-tsconfig-paths": "^5.1.4"
+		"pg": "8.18.0",
+		"react": "19.2.4",
+		"react-dom": "19.2.4",
+		"react-intersection-observer": "9.16.0",
+		"shadcn": "3.8.5",
+		"tailwind-merge": "3.4.1",
+		"tailwindcss": "4.1.18",
+		"tw-animate-css": "1.4.0",
+		"vite-tsconfig-paths": "5.1.4"
 	},
 	"devDependencies": {
-		"@prisma/client-runtime-utils": "7.4.0",
-		"@testing-library/dom": "^10.4.0",
-		"@testing-library/react": "^16.2.0",
-		"@types/node": "^22.10.2",
-		"@types/pg": "^8.11.10",
-		"@types/react": "^19.2.0",
-		"@types/react-dom": "^19.2.0",
-		"@vitejs/plugin-react": "^5.0.4",
-		"jsdom": "^27.0.0",
-		"prisma": "^7.4.0",
-		"typescript": "^5.7.2",
-		"vite": "^7.1.7",
-		"vitest": "^3.0.5",
-		"web-vitals": "^5.1.0"
-	}
+		"@tanstack/devtools-vite": "0.5.1",
+		"@testing-library/dom": "10.4.1",
+		"@testing-library/react": "16.3.2",
+		"@types/node": "22.19.11",
+		"@types/pg": "8.16.0",
+		"@types/react": "19.2.14",
+		"@types/react-dom": "19.2.3",
+		"@vitejs/plugin-react": "5.1.4",
+		"babel-plugin-react-compiler": "1.0.0",
+		"jsdom": "27.4.0",
+		"prisma": "7.4.0",
+		"typescript": "5.9.3",
+		"vite": "7.3.1",
+		"vitest": "3.2.4",
+		"web-vitals": "5.1.0"
+	},
+	"engines": {
+		"node": ">=22",
+		"bun": ">=1.3.5"
+	},
+	"packageManager": "bun@1.3.5"
 }
