@@ -15,7 +15,7 @@ const config = defineConfig({
 	},
 	plugins: [
 		devtools(),
-		nitro(),
+		nitro({ preset: 'bun' }),
 		// this is the plugin that enables path aliases
 		viteTsConfigPaths({
 			projects: ['./tsconfig.json'],
@@ -30,9 +30,6 @@ const config = defineConfig({
 
 		tailwindcss(),
 	],
-	ssr: {
-		noExternal: ['@prisma/client', /generated\/prisma/],
-	},
 });
 
 export default config;
